CHORE(client/admin): renamed AdmninPopup to CreateUserPopup
