+++ client/src/components/CreateUserPopup/RoleSpecificFields.tsx
@@ -1,6 +1,6 @@
 import { isVisibleToRole } from '../../utils/roles.utils';
 import { SelectField, InputField } from './Fields';
-import { RoleSpecificFieldsProps } from './AdminPopup.types';
+import { RoleSpecificFieldsProps } from './CreateUserPopup.types';
 
 const RoleSpecificFields = ({
   role,
